fix: avoid bucket usage package cycle
