+++ packages/scheduling/package.json
@@ -54,7 +54,6 @@
   },
   "dependencies": {
     "@alga-psa/authorization": "*",
-    "@alga-psa/billing": "*",
     "@alga-psa/clients": "*",
     "@alga-psa/core": "*",
     "@alga-psa/db": "*",
