+++ server/src/test/unit/billing/postDropRecurringObligationIdentity.test.ts
@@ -67,7 +67,7 @@ describe('post-drop recurring obligation identity', () => {
       '../packages/billing/src/actions/invoiceGeneration.ts',
       '../packages/billing/src/actions/recurringServicePeriodActions.ts',
       '../packages/billing/src/services/invoiceService.ts',
-      '../packages/billing/src/services/bucketUsageService.ts',
+      '../shared/billingClients/bucketUsageService.ts',
       '../packages/billing/src/lib/billing/billingEngine.ts',
     ].map((relativePath) => readFileSync(path.resolve(root, relativePath), 'utf8'));
