+++ server/src/test/unit/docs/servicePeriodFirstBillingPlan.contract.test.ts
@@ -136,10 +136,11 @@ const servicePeriodPostInventoryRefs = new Set([
   'packages/billing/src/components/billing-dashboard/AutomaticInvoices.tsx',
   'packages/billing/src/components/invoice-designer/inspector/TableEditorWidget.integration.test.tsx',
   'packages/billing/src/components/invoice-designer/inspector/widgets/TableEditorWidget.tsx',
-  'packages/billing/src/services/bucketUsageService.ts',
+  'shared/billingClients/bucketUsageService.ts',
   'packages/billing/tests/accounting/accountingExportInvoiceSelector.preview.test.ts',
   'packages/billing/tests/accounting/accountingExportValidation.rules.test.ts',
   'packages/billing/tests/automaticInvoices.groupedParentRows.test.tsx',
+  'packages/billing/tests/recurringApprovalBlockers.servicePeriodBoundary.test.ts',
   'packages/integrations/src/lib/xero/__tests__/xeroInvoiceMapping.test.ts',
   'server/src/lib/api/services/InvoiceService.ts',
   'server/src/test/infrastructure/billing/invoices/clientBillingCycleAnchors.test.ts',
